+++ code/packages/sdk/src/nodes/node.ts
@@ -25,7 +25,7 @@ import {
   mergeExceptArrays,
   moveNodeToContainerNT,
 } from '@/utils/utils';
-import { throttle } from 'lodash';
+import throttle from 'lodash/throttle';
 import type { KonvaEventObject } from 'konva/lib/Node';
 import { WEAVE_STAGE_DEFAULT_MODE } from './stage/constants';
 import { MOVE_TOOL_ACTION_NAME } from '@/actions/move-tool/constants';
@@ -1344,6 +1344,7 @@ export abstract class WeaveNode implements WeaveNodeBase {
     };
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   static getSchema(): z.ZodObject<any> {
     const baseNodeSchema = z.object({
       key: z